+++ named/server.go
@@ -9,11 +9,6 @@ type Server struct {
 	Name string
 }
 
-// NewServer returns a new Server.
-func NewServer(server serverz.Server, name string) *Server {
-	return &Server{server, name}
-}
-
 // GetName returns the name of a server.
 func (s *Server) GetName() string {
 	return s.Name