+++ CHANGELOG.md
@@ -11,6 +11,7 @@
 ### Removed
 
 - Unused logger interface
+- Unused constructors
 
 
 ## 0.9.0 - 2017-07-05